test(server): split the dispatch LLM-redaction tests into their own file (ARN-243)
Greptile's only note on #428 was that wasm_test.rs reached 588 lines. The
ADR-0166 callback-param tests are a self-contained group — which params survive
for a non-opted-in tenant and how their values are bounded — so they move to
llm_redaction_test.rs. 588 -> 409 + 186.
